+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## [Unreleased]
 ### Changed
 - fix(rpc): Set archival state in estimate tx expenses ([#3321](https://github.com/chainwayxyz/citrea/pull/3321))
+- fix(sequencer): Keep producing L2 blocks on stale DA tip ([#3324](https://github.com/chainwayxyz/citrea/pull/3324))
 
 ## [v2.7.0](2026-07-29)
 ### Changed